refactor: split httpApiV1 + consolidate moderation batches
