+++ convex/lib/reservedSlugs.ts
@@ -3,26 +3,25 @@ import type { MutationCtx, QueryCtx } from '../_generated/server'
 
 type ReservedSlug = Doc<'reservedSlugs'>
 
-export function pickLatestActiveReservation(reservations: ReservedSlug[]) {
-  const active = reservations.filter((r) => !r.releasedAt)
-  const latest = active.sort((a, b) => b.deletedAt - a.deletedAt)[0] ?? null
-  return { active, latest }
-}
+const DEFAULT_ACTIVE_LIMIT = 25
 
-export async function listReservedSlugsForSlug(
-  ctx: QueryCtx | MutationCtx,
-  slug: string,
-  limit = 10,
-) {
+function reservedSlugQuery(ctx: QueryCtx | MutationCtx, slug: string) {
   return ctx.db
     .query('reservedSlugs')
-    .withIndex('by_slug', (q) => q.eq('slug', slug))
-    .take(limit)
+    .withIndex('by_slug_active_deletedAt', (q) => q.eq('slug', slug).eq('releasedAt', undefined))
+    .order('desc')
+}
+
+export async function listActiveReservedSlugsForSlug(
+  ctx: QueryCtx | MutationCtx,
+  slug: string,
+  limit = DEFAULT_ACTIVE_LIMIT,
+) {
+  return reservedSlugQuery(ctx, slug).take(limit)
 }
 
 export async function getLatestActiveReservedSlug(ctx: QueryCtx | MutationCtx, slug: string) {
-  const reservations = await listReservedSlugsForSlug(ctx, slug)
-  return pickLatestActiveReservation(reservations).latest
+  return (await reservedSlugQuery(ctx, slug).take(1))[0] ?? null
 }
 
 export async function releaseDuplicateActiveReservations(
@@ -46,12 +45,12 @@ export async function reserveSlugForHardDeleteFinalize(
     expiresAt: number
   },
 ) {
-  const reservations = await listReservedSlugsForSlug(ctx, params.slug)
-  const { active, latest } = pickLatestActiveReservation(reservations)
+  const active = await listActiveReservedSlugsForSlug(ctx, params.slug)
+  const latest = active[0] ?? null
 
   if (latest) {
-    // Only extend the reservation if it matches the owner being deleted. If it points
-    // to someone else, it was likely created by reclaim and must not be overwritten.
+    // Only extend reservation if it matches the owner being deleted.
+    // If it points elsewhere, it likely came from a reclaim flow; do not overwrite.
     if (latest.originalOwnerUserId === params.originalOwnerUserId) {
       await ctx.db.patch(latest._id, {
         deletedAt: params.deletedAt,
@@ -82,8 +81,8 @@ export async function upsertReservedSlugForRightfulOwner(
     reason?: string
   },
 ) {
-  const reservations = await listReservedSlugsForSlug(ctx, params.slug)
-  const { active, latest } = pickLatestActiveReservation(reservations)
+  const active = await listActiveReservedSlugsForSlug(ctx, params.slug)
+  const latest = active[0] ?? null
 
   let keepId: Id<'reservedSlugs'>
   if (latest) {
@@ -112,8 +111,8 @@ export async function enforceReservedSlugCooldownForNewSkill(
   ctx: MutationCtx,
   params: { slug: string; userId: Id<'users'>; now: number },
 ) {
-  const reservations = await listReservedSlugsForSlug(ctx, params.slug)
-  const { active, latest } = pickLatestActiveReservation(reservations)
+  const active = await listActiveReservedSlugsForSlug(ctx, params.slug)
+  const latest = active[0] ?? null
   if (!latest) return
 
   if (latest.expiresAt > params.now && latest.originalOwnerUserId !== params.userId) {
@@ -123,9 +122,7 @@ export async function enforceReservedSlugCooldownForNewSkill(
     )
   }
 
-  // Original owner reclaiming, or reservation expired.
   await ctx.db.patch(latest._id, { releasedAt: params.now })
-
   await releaseDuplicateActiveReservations(ctx, active, latest._id, params.now)
 }
