+++ convex/skills.rateLimit.test.ts
@@ -87,7 +87,9 @@ describe('skills anti-spam guards', () => {
         if (table === 'reservedSlugs') {
           return {
             withIndex: (name: string) => {
-              if (name === 'by_slug') return { take: async () => [] }
+              if (name === 'by_slug_active_deletedAt') {
+                return { order: () => ({ take: async () => [] }) }
+              }
               throw new Error(`unexpected index ${name}`)
             },
           }
